Moved types.rs to c_types.rs so types can be used for higher level abstractions
