+++ src/bindings.rs
@@ -1,6 +1,6 @@
 #![allow(non_camel_case_types, non_upper_case_globals, non_snake_case, improper_ctypes)]
 
-use types;
+use c_types;
 
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
