+++ src/error.rs
@@ -1,16 +1,16 @@
 use bindings;
-use types;
+use c_types;
 
-pub struct Error(types::c_int);
+pub struct Error(c_types::c_int);
 
 impl Error {
     pub const EINVAL: Self = Error(-(bindings::EINVAL as i32));
 
-    pub fn from_kernel_errno(errno: types::c_int) -> Error {
+    pub fn from_kernel_errno(errno: c_types::c_int) -> Error {
         return Error(errno);
     }
 
-    pub fn to_kernel_errno(&self) -> types::c_int {
+    pub fn to_kernel_errno(&self) -> c_types::c_int {
         return self.0;
     }
 }
